pyproject.toml
[tool.poetry]
name = "tcnchatbot"
version = "0.1.0"
description = "A complex moderation bot tailored for a specific group."
authors = ["symmetricalboy <email@symmetricalboy.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
python-telegram-bot = {version = "^22.6", extras = ["webhooks", "job-queue"]}
python-dotenv = "^1.0.1"
asyncpg = "^0.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
